+++ server/src/modules/leads/lead.constants.js
@@ -1,0 +1,7 @@
+const LEAD_STAGES = ['New', 'Contacted', 'Qualified', 'Won', 'Lost']
+const DEFAULT_LEAD_STAGE = 'New'
+
+module.exports = {
+  DEFAULT_LEAD_STAGE,
+  LEAD_STAGES,
+}