+++ server/src/routes/index.js
@@ -1,10 +1,12 @@
 const { Router } = require('express')
 const authRoutes = require('../modules/auth/auth.routes')
+const leadsRoutes = require('../modules/leads/leads.routes')
 const healthRoutes = require('./health.routes')
 
 const router = Router()
 
 router.use(healthRoutes)
 router.use('/auth', authRoutes)
+router.use('/leads', leadsRoutes)
 
 module.exports = router